mcp-servers/google-calendar/requirements.txt
mcp[cli]>=1.2.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
uvicorn>=0.32.0
